+++ lib/input/constructor.go
@@ -235,6 +235,8 @@ func (c *Config) UnmarshalJSON(bytes []byte) error {
 			return fmt.Errorf("failed to parse plugin config: %v", err)
 		}
 		aliased.Plugin = dummy.Conf
+	} else {
+		aliased.Plugin = nil
 	}
 
 	*c = Config(aliased)
@@ -262,6 +264,8 @@ func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 			return err
 		}
 		aliased.Plugin = conf
+	} else {
+		aliased.Plugin = nil
 	}
 
 	*c = Config(aliased)