+++ lib/output/constructor.go
@@ -108,7 +108,7 @@ type Config struct {
 	Kinesis       writer.KinesisConfig       `json:"kinesis" yaml:"kinesis"`
 	MQTT          writer.MQTTConfig          `json:"mqtt" yaml:"mqtt"`
 	Nanomsg       NanomsgConfig              `json:"nanomsg" yaml:"nanomsg"`
-	NATS          NATSConfig                 `json:"nats" yaml:"nats"`
+	NATS          writer.NATSConfig          `json:"nats" yaml:"nats"`
 	NATSStream    NATSStreamConfig           `json:"nats_stream" yaml:"nats_stream"`
 	NSQ           NSQConfig                  `json:"nsq" yaml:"nsq"`
 	Plugin        interface{}                `json:"plugin,omitempty" yaml:"plugin,omitempty"`
@@ -143,7 +143,7 @@ func NewConfig() Config {
 		Kinesis:       writer.NewKinesisConfig(),
 		MQTT:          writer.NewMQTTConfig(),
 		Nanomsg:       NewNanomsgConfig(),
-		NATS:          NewNATSConfig(),
+		NATS:          writer.NewNATSConfig(),
 		NATSStream:    NewNATSStreamConfig(),
 		NSQ:           NewNSQConfig(),
 		Plugin:        nil,
@@ -237,6 +237,8 @@ func (c *Config) UnmarshalJSON(bytes []byte) error {
 			return fmt.Errorf("failed to parse plugin config: %v", err)
 		}
 		aliased.Plugin = dummy.Conf
+	} else {
+		aliased.Plugin = nil
 	}
 
 	*c = Config(aliased)
@@ -264,6 +266,8 @@ func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 			return err
 		}
 		aliased.Plugin = conf
+	} else {
+		aliased.Plugin = nil
 	}
 
 	*c = Config(aliased)